+++ src/Board.h
@@ -9,12 +9,12 @@
 class Board
 {
 public:
-    Board(const unsigned int& size)
+    Board(const size_t& size)
         : m_size(size),
           m_board(size, std::vector<std::string>(size))
     {}
 
-    typedef std::pair<unsigned int, unsigned int> BoardPosition;
+    typedef std::pair<size_t, size_t> BoardPosition;
     typedef std::vector<BoardPosition> BoardPositions;
   
     void updateAndPrint(const BoardPositions& positions);
@@ -24,9 +24,9 @@ public:
 
 private:
     std::string invertColors(const std::string& c) const;
-    void updatePosition(const unsigned int& index_row, const unsigned int& index_col, const std::string& c); 
+    void updatePosition(const size_t& index_row, const size_t& index_col, const std::string& c); 
 
-    const unsigned int m_size;
+    const size_t m_size;
     std::vector<std::vector<std::string>> m_board;
 
 };
